initial version
The script can now check if any mails has been received on previous day/yesterday in a folder @my_folder under Inbox under the mailbox defined in @my_mailbox. If mail has been received, script will exit. If not, script assumes server is not up or has some issue and mails were not sent out. Server needs to be checked on immediately. Script will display a popup with warning in this case and exit only when user closes the popup.